Refactor : 속성명 수정 및 파일 위치 변경
