+++ src/main/java/com/odos/odos_server/Enum/Feeling.java
@@ -1,6 +1,6 @@
 package com.odos.odos_server.Enum;
 
-public enum Emotion {
+public enum Feeling {
   SAD, // 슬픔
   NORMAL, // 보통
   HAPPY, // 기쁨
