+++ src/main/java/com/odos/odos_server/Enum/Job.java
@@ -2,5 +2,5 @@ package com.odos.odos_server.Enum;
 
 public enum Job {
   STUDENT,
-  EMPLOYEE
+  WORKER
 }
