+++ src/main/java/com/odos/odos_server/Enum/MemberChallengeRole.java
@@ -1,0 +1,8 @@
+package com.odos.odos_server.Enum;
+
+public enum MemberChallengeRole {
+    HOST,
+    APPLICANT,
+    REQUESTED,
+    NONE
+}
