+++ src/main/java/com/odos/odos_server/Enum/ReportType.java
@@ -1,7 +1,7 @@
 package com.odos.odos_server.Enum;
 
 public enum ReportType {
-  NOT_GOOD_TITLE_CONTENT, // 적절하지 않은 제목과 내용
-  NOT_GOOD_PICTURE, // 적잘하지 않은 사진
+  BAD_TITLE_CONTENT, // 적절하지 않은 제목과 내용
+  BAD_IMAGE, // 적잘하지 않은 사진
   ETC // 기타
 }
