+++ src/main/java/com/odos/odos_server/diary/DiaryGoal.java
@@ -3,7 +3,7 @@ package com.odos.odos_server.diary;
 // 챌린지 목표-일지
 
 import com.odos.odos_server.challenge.ChallengeGoal;
-import com.odos.odos_server.member.MemberChallenge;
+import com.odos.odos_server.challenge.MemberChallenge;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
